Make managed-notifications play well with multiple sources for the same bucket (#2710)
The Panther notifications feature didn't really account for multiple s3 sources for the same bucket. This scenario
would make the feature unusable if some s3 sources would have the same bucket and overlapping prefixes.

Although this might not seem a common scenario, it is currently allowed and we should support it.